feat: adding a new page for reveal identity
