--- packages/demo/src/components/GetCommitment/index.ts
@@ -1,3 +1,0 @@
-import { GetCommitment } from "./GetCommitment";
-
-export default GetCommitment;
