+++ packages/demo/src/constants/paths.ts
@@ -3,5 +3,6 @@ export enum Paths {
     HOME = "/home",
     CONNECT = "/connect/:isChangeIdentityParam",
     GET_IDENTITY_METADATA = "/get-connected-metadata",
-    IMPORT_IDENTITY =  "/import-identity"
+    IMPORT_IDENTITY =  "/import-identity",
+    REVEAL_IDENTITY_COMMITMENT = "/reveal-identity-commitment"
 }
