+++ packages/demo/src/pages/RevealIdentity/index.ts
@@ -1,0 +1,3 @@
+import {lazy} from "react";
+
+export default lazy(() => import("./RevealIdentity"));
